% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotenuse.R
\name{hypotenuse}
\alias{hypotenuse}
\title{Calculate the hypotenuse for an N-dimensional polygon}
\usage{
hypotenuse(...)
}
\arguments{
\item{...}{Dimension lengths}
}
\description{
Calculate the hypotenuse for an N-dimensional polygon
}
